set up ui

+++ mobile/app/src/main/java/com/example/chatapp/model/response/UserResponse.kt
@@ -1,11 +1,10 @@
 package com.example.chatapp.model.response
 
 data class UserResponse (
-    val id: Long? = null,
+    val id: Int? = null,
     val username: String? = null,
     val email: String? = null,
     val profilePicUrl: String? = null,
     val createdAt: String? = null,
     val updatedAt: String? = null
 )
-
